Update total count alg
