+++ src/PageCharacter/CharacterDisplay/Tabs/TabOptimize/Components/BuildAlert.tsx
@@ -1,8 +1,17 @@
 import { Alert, Grid, LinearProgress, styled, Typography } from '@mui/material';
-import React, { ReactNode } from 'react';
+import { ReactNode } from 'react';
 import { timeStringMs } from '../../../../../Util/TimeUtil';
 
 export const warningBuildNumber = 10000000
+export type BuildStatus = {
+  type: "active" | "inactive"
+  tested: number      // tested, including `failed`
+  failed: number      // tested but fail the filter criteria, e.g., not enough EM
+  skipped: number
+  total: number
+  startTime?: number
+  finishTime?: number
+}
 
 const Monospace = styled("strong")({
   fontFamily: "monospace"
@@ -12,13 +21,17 @@ const BorderLinearProgress = styled(LinearProgress)(({ theme }) => ({
   height: 10,
   borderRadius: 5,
 }));
-export default function BuildAlert({ totBuildNumber, generatingBuilds, generationSkipped, generationProgress, generationDuration, characterName, maxBuildsToShow }) {
-  const totalBuildNumberString = totBuildNumber?.toLocaleString() ?? totBuildNumber
-  const totalUnskipped = totBuildNumber - generationSkipped
-  const generationProgressString = generationProgress?.toLocaleString() ?? generationProgress
-  const generationSkippedString = generationSkipped?.toLocaleString() ?? generationSkipped
-  const totalUnskippedString = totalUnskipped?.toLocaleString() ?? totalUnskipped
-  const generationSkippedText = !!generationSkipped && <span>(<b>{generationSkippedString}</b> skipped)</span>
+export default function BuildAlert({ status: { type, tested, failed, skipped, total, startTime, finishTime }, characterName }: { status: BuildStatus, characterName: Displayable }) {
+  const hasTotal = isFinite(total)
+
+  const generatingBuilds = type !== "inactive"
+  const unskipped = total - skipped
+
+  const testedString = <Monospace>{tested.toLocaleString()}</Monospace>
+  const unskippedString = <Monospace>{unskipped.toLocaleString()}</Monospace>
+  const skippedText = !!skipped && <span>(<b>{<Monospace>{skipped.toLocaleString()}</Monospace>}</b> skipped)</span>
+
+  const durationString = <Monospace>{timeStringMs(Math.round((finishTime ?? performance.now()) - (startTime ?? NaN)))}</Monospace>
 
   let color = "success" as "success" | "warning" | "error"
   let title = "" as ReactNode
@@ -26,22 +39,15 @@ export default function BuildAlert({ totBuildNumber, generatingBuilds, generatio
   let progress = undefined as undefined | number
 
   if (generatingBuilds) {
-    progress = generationProgress * 100 / (totalUnskipped)
-    title = <Typography>Generating and testing <Monospace>{generationProgressString}/{totalUnskippedString}</Monospace> build configurations against the criteria for <b>{characterName}</b>. {generationSkippedText}</Typography>
-    subtitle = <Typography>Time elapsed: <Monospace>{timeStringMs(Math.round(generationDuration))}</Monospace></Typography>
-  } else if (!generatingBuilds && generationProgress) {//done
+    progress = tested * 100 / (unskipped)
+    title = <Typography>Generating and testing {testedString}{hasTotal ? <>/{unskippedString}</> : undefined} build configurations against the criteria for <b>{characterName}</b>. {skippedText}</Typography>
+    subtitle = <Typography>Time elapsed: {durationString}</Typography>
+  } else if (tested + skipped) {
     progress = 100
-    title = <Typography>Generated and tested <Monospace>{totalUnskippedString}</Monospace> Build configurations against the criteria for <b>{characterName}</b>. {generationSkippedText}</Typography>
-    subtitle = <Typography>Total duration: <Monospace>{timeStringMs(Math.round(generationDuration))}</Monospace></Typography>
+    title = <Typography>Generated and tested {testedString} Build configurations against the criteria for <b>{characterName}</b>. {skippedText}</Typography>
+    subtitle = <Typography>Total duration: {durationString}</Typography>
   } else {
-    if (totBuildNumber === 0) {
-      title = <Typography>Current configuration will not generate any builds for <b>{characterName}</b>. Please change your Artifact configurations, or add/include more Artifacts.</Typography>
-      color = "error"
-    } else if (totBuildNumber > warningBuildNumber) {
-      title = <Typography>Current configuration will generate <Monospace>{totalBuildNumberString}</Monospace> potential builds for <b>{characterName}</b>. This might take quite a while to generate...</Typography>
-      color = "warning"
-    } else
-      title = <Typography>Current configuration {totBuildNumber <= maxBuildsToShow ? "generated" : "will generate"} <Monospace>{totalBuildNumberString}</Monospace> builds for <b>{characterName}</b>.</Typography>
+    title = <Typography>Current configuration will generate builds for <b>{characterName}</b>.</Typography>
   }
 
   return <Alert severity={color} variant="filled" sx={{
@@ -49,14 +55,14 @@ export default function BuildAlert({ totBuildNumber, generatingBuilds, generatio
       flexGrow: 1
     }
   }}>
-    {title && title}
-    {subtitle && subtitle}
+    {title}
+    {subtitle}
     {progress !== undefined && <Grid container spacing={1} alignItems="center">
-      <Grid item>
+      {hasTotal && <Grid item>
         <Typography>{`${progress.toFixed(1)}%`}</Typography>
-      </Grid>
+      </Grid>}
       <Grid item flexGrow={1} >
-        <BorderLinearProgress variant="determinate" value={progress} color="primary" />
+        <BorderLinearProgress variant={hasTotal ? "determinate" : "indeterminate"} value={progress} color="primary" />
       </Grid>
     </Grid>}
   </Alert>
